requirements.txt
Django
psycopg2-binary
redis
djangorestframework
djangorestframework-simplejwt
python-decouple
drf-yasg
django-cors-headers
django-ratelimit
gunicorn
django-redis
dj-database-url
whitenoise